twister: cleanup imports and move size classes out
More python import optimizations and cleanup and move size_calc class
out into own file.
